fix: Don't overwrite symlinks
Second attempt at fixing #3.

Since a3527b8c, wrkn writes files atomically. Since it doesn't check if the 
data file is a symlink, a symlink is overwritten rather than dereferenced
and the target written to.

This change checks if the data file is a symlink, and if it is, resolves its
target, and uses the resolved path to write to.

The change also adds an integration regression test for this behaviour.

Closes #3